requirements.txt
blinker==1.9.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
comtypes==1.4.12
cryptography==46.0.3
Flask==2.3.2
flask-cors==6.0.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.2
lxml==6.0.2
MarkupSafe==3.0.2
nltk==3.9.2
numpy==2.3.5
pdfminer.six==20251107
pycparser==2.23
pyjokes==0.8.3
pypiwin32==223
python-docx==1.2.0
pyttsx3==2.99
pywin32==311
regex==2025.11.3
requests==2.32.5
scikit-learn==1.7.2
scipy==1.16.3
threadpoolctl==3.6.0
tqdm==4.67.1
typing_extensions==4.15.0
urllib3==2.5.0
Werkzeug==3.1.3
